requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
boto3==1.28.57
botocore==1.31.57
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0

# Machine Learning (optional - install if using advanced features)
scikit-learn==1.3.0
# tensorflow==2.13.0  # Uncomment if using TensorFlow models
# torch==2.0.1  # Uncomment if using PyTorch models
# torchvision==0.15.2  # Uncomment if using PyTorch vision models
# ultralytics==8.0.0  # Uncomment if using YOLOv8 models
# mediapipe==0.10.0  # Uncomment if using MediaPipe models

# Image and Video Processing
imageio==2.31.1
moviepy==1.0.3
ffmpeg-python==0.2.0

# Data Processing
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2

# Development and Testing
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
pre-commit==3.3.3

# Production and Monitoring
redis==4.6.0
watchdog==3.0.0
psutil==5.9.5

# Cloud Services (optional)
# azure-storage-blob==12.17.0  # Uncomment if using Azure
# google-cloud-storage==2.10.0  # Uncomment if using Google Cloud
